fix(bugfix): 자잘한 버그픽스 및 리팩토링
