+++ hooks/api/auth/index.ts
@@ -1,5 +1,7 @@
+import { useRouter } from "next/router";
 import { useQueryClient } from "react-query";
 import { postAuth, postAuthAnonymous, postLogout } from "~/api/auth";
+import { Toast } from "~/components/uis";
 import { queryKeys } from "~/consts/react-query";
 import { useCoreMutation } from "~/hooks/api/core";
 
@@ -20,9 +22,19 @@ export const usePostAuthAnonymousMutation = () => {
 };
 
 export const usePostLogoutMutation = () => {
+  const router = useRouter();
   const queryClient = useQueryClient();
 
   return useCoreMutation(postLogout, {
-    onSuccess: () => queryClient.invalidateQueries(queryKeys.auth),
+    onSuccess: () => {
+      Toast.show("로그아웃에 성공하였습니다.", {
+        duration: 3000,
+      });
+
+      localStorage.clear();
+      queryClient.invalidateQueries(queryKeys.auth);
+
+      router.replace("/");
+    },
   });
 };
